+++ spec/unit/windows_spec.rb
@@ -21,8 +21,7 @@ describe 'java_se::default' do
 
     it 'installs java' do
       expect(chef_run).to run_execute(
-        "install jdk-#{VERSION}_windows-x64_bin.exe to C:\\Program Files\\Java\\jdk-#{VERSION}"\
-                                      " with JRE C:\\Program Files\\Java\\jre-#{VERSION}"
+        "install jdk-#{VERSION}_windows-x64_bin.exe to C:\\Program Files\\Java\\jdk-#{VERSION}"
       )
     end
 
@@ -30,13 +29,8 @@ describe 'java_se::default' do
       expect(chef_run).to create_env('JAVA_HOME')
     end
 
-    it 'sets JRE_HOME' do
-      expect(chef_run).to create_env('JRE_HOME')
-    end
-
     it 'sets PATH' do
       expect(chef_run).to modify_env('Add java_se JDK to path').with(key_name: 'PATH')
-      expect(chef_run).to modify_env('Add java_se JRE to path').with(key_name: 'PATH')
     end
 
     it 'creates dir' do